Fixed major issue with header internal data alignment
This issue was with header internal data during the write.
When the data was getting committed to disk via NetCDF4 functions,
the counter was getting incremented right before the NetCDF4
call. While this did not cause any major issues with similar data
points between, it did cause issues with data towards the end.
Because the counter was one position ahead, it caused data added
towards the end to not get added at all, for all types of variables.

This was fixed by swapping the lines - having the counter get
incremented AFTER theh nf90_put_att call, not before.