Overhauled header code to just be a simple wrapper for NetCDF4
Removed all caching code for the header part, since it's just modifying
the NetCDF4 global attributes, which are meant for overwriting. Since
it's much simpler (no data storage needed, no appending, etc.), the
caching code was scrapped in favor of the header portion becoming a
thin wrapper for the NetCDF4 nf90_put_att function. This allows the
header writing to be MUCH faster, due to removing unneeded
functionality.