Fixed major bug with reallocating index arrays
Fixed a major bug where additional data will reallocate the index
arrays with significant growth, due to the reallocator using and
incrementing the metadata multiplier for every variable reallocated.
With a lot of variables, this causes the multiplier to grow
significantly, resulting in significant memory consumption and possibly
a crash. This was fixed by creating a per-variable multiplier, and
using that instead. This may be fixed in a future using a scalar,
but it will require some code reworking to ensure this doesn't
happen again!