Added initial buffering support for chaninfo
Added buffering support for the chaninfo type, and laid out initial
support for buffering for other variables. Buffering is useful when
memory usage is a concern. This allows variables to be saved into
NetCDF's internal storage at an earlier period in time so that the
existing NCLayer memory can be reused without needing to expand,
allowing memory usage to be constrained. This requires definition
locking in order to ensure that the variables are ready to be
written to NetCDF.

One other change - chaninfo data writing now checks if there is any
data to be written before attempting, due to the way that buffered
writing is set up. As a plus, this speeds up writing when there are
no additional data to be written.