Fix serious issues with data2d storage
Fortran is column based, and NetCDF4 respects that order. As a result,
data MUST be stored in the array in columns of data in order to
translate into "rows" of data. Previously, allocation of the temporary
2D array and storage of said array used rows (e.g. (j, :) vs. the
correct (:, j) indexing), causing data to appear out of order. This has
now been fixed.

Additionally, issues with overfilling data have been encountered. If
the subarray of data stored is less than the maximum length (e.g.
an array of (/ 2, 3 /) was added to a fuller array with (/ 2, 3, 4 /)
sized arrays, or a length of 2 vs. 3), the empty spot would be filled
with -1s (compiler [un]defined behavior?), resulting in the negation
of the empty fill.

Finally, there were issues with overfilling the array and exceeding
the bounds. The index array was added to the length array without
subtracting one, resulting in a attempt to store incorrectly a larger
array (LENGTH + 1), as opposed to the correct length.

The last two issues have been fixed by specifying absolute indexing
and changing the ending data array index from (INDEX + LENGTH) to
(INDEX + LENGTH - 1).

data2d storage should (hopefully) work now!