Added strict bound checking to chaninfo, metadata, and data2d
With chaninfo, checks were already in place to detect whether the
provided data filled the nchans data width or not. If it didn't,
it displayed a warning. With the new strict bound checking option,
failing this check can now result in an actual error when strict
checking is turned on. Of course, strict checking is disabled
by default, and warnings can still be emitted.

Strict bounds checking can be toggled on and off globally via
nc_diag_set_strict(), or by data type with
nc_diag_chaninfo_set_strict(), nc_diag_metadata_set_strict(), and/or
nc_diag_data2d_set_strict(). test_netcdf_layer.f90 provides a one-lin
example of the usage - just a boolean!

With metadata, a new check is implemented to see whether all of the
variables have the same amount of nobs. For instance, if variable
"asdf" has 10 entries in the nobs dimension (or just 10 entries),
"ghjk" should also have 10 entries as well. If "ghjk" has 9 or fewer
entries, this will result in a warning. Combined with the strict bound
checking option, this can also result in an error as well if strict
bound checking is enabled.

With data2d, two new checks are implementd to ensure data consistency
as well. Like with metadata, a new check is implemented to see whether
all of the variables have the same amount of obs. In addition, another
check is implemented to ensure that each subarray has the same amount
of elements. For instance, in variable "asdf", a subarray entry along
the nobs dimension could be (/ 1, 2, 3 /), with 3 entries. In that
case, all of the other entries within that variable should have the
same amount of elements, e.g. (/ 4, 5, 6 /), or 3 entries. If any other
entries have 2 or less entries, e.g. (/ 7, 8 /), a warning will be
displayed notifying about inconsistencies. Fewer values will result
in the last missing value(s) being filled with the NetCDF4 "missing"
entry. Note that adding an entry with more than 3 entries won't error
here, due to the maximum length checks implemented within the
definition subroutine and the data entry subroutins.

These checks are enabled to ensure that the user knows if there are any
inconsistencies with the data.

Combined with the strict bound checking options, these new checks can
also result in error(s) as well if strict bound checking is enabled.