Fix data2d string writing to use temp var for allocation
The Fortran compiler optimized itself to use the allocated array
directly, which in many cases would cause runtime errors to occur,
even when the array wasn't going to be used as all (such as a
non-string variable). Therefore, the length was fetched from the
allocated array into a temporary variable, and then said temporary
variable is used for allocating the temporary array.