Improved chunking for data2d and metadata
Fixed metadata chunking so that the chunking is much more
efficient and easier to handle. Chunking for strings used to
use 1 as a chunking dimension, which was inefficient. It now
uses the maximum string length for that variable, since that
length is how the strings are typically stored and accessed.

The 1024 is also replaced by a global parameter for easier
changing. The new parameter, NLAYER_CHUNKING, has also been
updated with a value of 16384 for better compression and
performance.

As a result of better understanding NetCDF chunking, data2d
chunking has also been enabled, using the maximum string
length as a chunking parameter when applicable, as well as
using the maximum array length for all data2d variables.
For the unlimited dimension, we also use the new global
NLAYER_CHUNKING parameter for easier updating.

These major chunking updates should hopefully increase
performance in reading/writing the NetCDF4 file, and
hopefully reduce the resulting file size as well.