Fix chaninfo append for string variables
max_str_lens was not being set with chaninfo, so the string
bounds check fails when definitions are locked (which is
always the case for appending). This was fixed by setting
the max_str_lens to the input NetCDF file's variable string
dimension. Since the max_str_lens for the variable is now
defined, constraint checks now work on append.