Fixed ncdr_dims to use relative indexing (instead of actual NCIDs)
ncdr_dims used to use the old system of NCIDs (which didn't work).
As a result, things broke when the dimension APIs were used. This is
fixed by wrapping the NCIDs with nc_diag_read_get_index_from_ncid(),
which is done for ncdr_vars and other portions of nc_diag_read.