Modularized and improved Makefiles, created new static library
Modularized Makefiles to split build configuration into seperate
Makefile configurations. This allows Makefiles to become shorter and
simpler. We also added a rule to build a static library, which will
allow for easier linking and inclusion for developers. In addition,
we also created a Makefile.libconfig for developers to use to make it
easier to include nc_diag_read in their programs.