Migrate nc_diag_cat to modular source/building
Previously, nc_diag_cat was built using #includes. The main source
file would include all other dependencies, and after pre-processing
the entire source by the compiler, the result would be one big file.
This is really messy and isn't good for organizing things.

As a result, the file has been "split up" into different logical
modules. All of the modules will still be compiled together to
the nc_diag_cat.x program, but now the builds are incremental, and
the code is (hopefully!) easier to manage!